+++ tests_and_analysis/test/euphonic_test/test_spectrum1dcollection.py
@@ -7,11 +7,13 @@ import pytest
 
 from euphonic import ureg
 from euphonic.spectra import Spectrum1DCollection
-from tests_and_analysis.test.utils import get_data_path, get_castep_path
-
-from .test_spectrum1d import (get_spectrum1d, get_expected_spectrum1d,
-                              check_spectrum1d, check_unit_conversion,
-                              check_property_setters)
+from tests_and_analysis.test.utils import (
+    get_data_path, get_castep_path, get_spectrum_from_text,
+    check_spectrum_text_header)
+from tests_and_analysis.test.euphonic_test.test_spectrum1d import (
+    get_spectrum1d, get_expected_spectrum1d,
+    check_spectrum1d, check_unit_conversion,
+    check_property_setters)
 
 
 class ExpectedSpectrum1DCollection:
@@ -78,16 +80,16 @@ def get_spectrum_dir():
     return os.path.join(get_data_path(), 'spectrum1dcollection')
 
 
-def get_json_file(json_filename):
-    return os.path.join(get_spectrum_dir(), json_filename)
+def get_full_path(filename):
+    return os.path.join(get_spectrum_dir(), filename)
 
 
 def get_spectrum1dcollection(json_filename):
-    return Spectrum1DCollection.from_json_file(get_json_file(json_filename))
+    return Spectrum1DCollection.from_json_file(get_full_path(json_filename))
 
 
 def get_expected_spectrum1dcollection(json_filename):
-    return ExpectedSpectrum1DCollection(get_json_file(json_filename))
+    return ExpectedSpectrum1DCollection(get_full_path(json_filename))
 
 
 def check_spectrum1dcollection(actual_spectrum, expected_spectrum):
@@ -135,7 +137,7 @@ class TestSpectrum1DCollectionCreation:
         json_file = request.param
         expected_spectrum = get_expected_spectrum1dcollection(json_file)
         spectrum = Spectrum1DCollection.from_json_file(
-            get_json_file(json_file))
+            get_full_path(json_file))
         return spectrum, expected_spectrum
 
     @pytest.fixture(params=[
@@ -271,38 +273,52 @@ class TestSpectrum1DCollectionCreation:
 @pytest.mark.unit
 class TestSpectrum1DCollectionSerialisation:
 
-    @pytest.fixture(params=[
+    # Note that when writing .text there must be the same number of
+    # x_data and y_data points so bin centres will be used, this and
+    # using fmt may mean the output spectrum is slightly different.
+    # x_tick_labels will also be lost
+    @pytest.mark.parametrize('in_json, out_json, kwargs', [
+        ('quartz_666_coh_pdos.json',
+         'quartz_666_coh_pdos_from_text_file_fmt.json',
+         {'fmt': '%.2f'}),
+        ('gan_bands.json', 'gan_bands_from_text_file.json', {}),
+        ('methane_pdos.json', 'methane_pdos_from_text_file_fmt.json',
+         {'fmt': ['%.4f'] + ['%.2f']*5})])
+    def test_serialise_to_text_file(self, in_json, out_json, kwargs, tmpdir):
+        spectrum = get_spectrum1dcollection(in_json)
+        # Serialise
+        output_file = str(tmpdir.join('tmp.test'))
+        spectrum.to_text_file(output_file, **kwargs)
+        # Deserialise
+        deserialised_spectrum = get_spectrum_from_text(
+            output_file, is_collection=True)
+        expected_deserialised_spectrum = get_spectrum1dcollection(out_json)
+        check_spectrum_text_header(output_file)
+        check_spectrum1dcollection(
+            expected_deserialised_spectrum, deserialised_spectrum)
+
+    @pytest.mark.parametrize('spectrum', [
         get_spectrum1dcollection('gan_bands.json'),
         get_spectrum1dcollection('methane_pdos.json')])
-    def serialise_to_json_file(self, request, tmpdir):
-        spectrum = request.param
+    def test_serialise_to_json_file(self, spectrum, tmpdir):
         # Serialise
         output_file = str(tmpdir.join('tmp.test'))
         spectrum.to_json_file(output_file)
         # Deserialise
         deserialised_spectrum = (Spectrum1DCollection
                                  .from_json_file(output_file))
-        return spectrum, deserialised_spectrum
-
-    def test_serialise_to_file(self, serialise_to_json_file):
-        spectrum, deserialised_spectrum = serialise_to_json_file
         check_spectrum1dcollection(spectrum, deserialised_spectrum)
 
-    @pytest.fixture(params=[
+    @pytest.mark.parametrize('json_file', [
         'gan_bands.json',
         'methane_pdos.json'])
-    def serialise_to_dict(self, request):
-        json_file = request.param
+    def test_serialise_to_dict(self, json_file):
         spectrum = get_spectrum1dcollection(json_file)
         expected_spectrum = get_expected_spectrum1dcollection(json_file)
         # Convert to dict, then back to object to test
         spectrum_dict = spectrum.to_dict()
         spectrum_from_dict = Spectrum1DCollection.from_dict(spectrum_dict)
-        return spectrum_from_dict, expected_spectrum
-
-    def test_serialise_to_dict(self, serialise_to_dict):
-        spectrum, expected_spectrum = serialise_to_dict
-        check_spectrum1dcollection(spectrum, expected_spectrum)
+        check_spectrum1dcollection(spectrum_from_dict, expected_spectrum)
 
 
 @pytest.mark.unit